requirements.txt
python-telegram-bot==20.7
google-genai==0.3.0
redis==5.0.1
